feat[6.21-22]: use react query for creating and voting anecdotes
