+++ part6/query-anecdotes/src/App.jsx
@@ -1,19 +1,25 @@
-import { useQuery } from '@tanstack/react-query'
+import { useQuery,  useMutation, useQueryClient  } from '@tanstack/react-query'
+
+import anecdoteService from './services/Anecdotes'
 
 import AnecdoteForm from './components/AnecdoteForm'
 import Notification from './components/Notification'
 
-import anecdoteService from './services/Anecdotes'
-
 const App = () => {
+
+  const queryClient = useQueryClient()
+
+  const updatedAnecdoteMutation = useMutation(anecdoteService.updateExisting, {
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ['anecdotes'] })
+    }
+  })
 
   const result = useQuery({
     queryKey: ['anecdotes'],
     queryFn: anecdoteService.getAll,
     retry: 1
   })
-
-  console.log(JSON.parse(JSON.stringify(result)))
 
   if ( result.isLoading ) {
     return <div>loading data...</div>
@@ -23,9 +29,7 @@ const App = () => {
     return <div>anecdote service not available due to problems in server</div>
   }
 
-  const handleVote = (anecdote) => {
-    console.log('vote')
-  }
+  const handleVote = (anecdote) => updatedAnecdoteMutation.mutate({ ...anecdote, votes: anecdote.votes + 1 })
 
   const anecdotes = result.data
